requirements.txt
googlemaps
requests
crawl4ai
python-dotenv
openai